added CleanUp: Pointer-Get deduplication
